Moved images and begun migration of gamedev and minecraft pages
