+++ lib/MCScreenshots.ts
@@ -1,9 +1,9 @@
-import LucHouse from "../img/MCScreenShots/LucHouse.png";
-import SpawnBridge from "../img/MCScreenShots/SpawnBridge.png";
-import TheDarkThiefPowerStation from "../img/MCScreenShots/TheDarkThiefPowerStation.png";
-import AFactory from "../img/MCScreenShots/AFactory.png";
-import TheDarkThiefHome from "../img/MCScreenShots/TheDarkThiefHobbitHome.png";
-import OpenSubmission from "../img/MCScreenShots/SubmitYOurScreenshots.png";
+import LucHouse from "../public/img/MCScreenShots/LucHouse.png";
+import SpawnBridge from "../public/img/MCScreenShots/SpawnBridge.png";
+import TheDarkThiefPowerStation from "../public/img/MCScreenShots/TheDarkThiefPowerStation.png";
+import AFactory from "../public/img/MCScreenShots/AFactory.png";
+import TheDarkThiefHome from "../public/img/MCScreenShots/TheDarkThiefHobbitHome.png";
+import OpenSubmission from "../public/img/MCScreenShots/SubmitYOurScreenshots.png";
 
 const mcImages = {
     LucHouse: LucHouse,
