+++ package.json
@@ -23,7 +23,7 @@
         "@chakra-ui/react": "^2.8.0",
         "@emotion/react": "^11.11.1",
         "@emotion/styled": "^11.11.0",
-        "framer-motion": "^10.16.1",
+        "framer-motion": "^11.0.8",
         "ioredis": "^5.3.2",
         "mariadb": "^3.2.1",
         "ncp": "^2.0.0",
