+++ pages/index.tsx
@@ -4,8 +4,8 @@ import PageCard from "../components/PageCard/PageCard";
 import { Box, Flex, Heading, Stack, Text, VStack } from "@chakra-ui/react";
 import NextLink from "next/link";
 import Image from "next/image";
-import bannerImg from "../img/Homepage-Background.svg";
-import bannerImgCover from "../img/Homepage-Background-Text.svg";
+import bannerImg from "../public/img/Homepage-Background.svg";
+import bannerImgCover from "../public/img/Homepage-Background-Text.svg";
 import { GetServerSideProps } from "next";
 import { getNewsCards, NewsCard } from "../lib/News";
 import { EventCard, getEventCards } from "../lib/Event";
